Refactor timestamp clone cleanup logic into shared abstraction (Fixes #58)
